Implement Ui, Storage, Parser and TaskList Classes
Implemented more classes to abstract out all the functionalities further.

Abstracting into these classes will improve code readability, ease of debugging and reduce code complexity.

Let's
* Implement Ui, Storage, Parser, TaskList Classes
* Change Alfred class to now use these subclasses
